Made main generator a class
